refactor: 역할에 따른 숙소 관련 api 분리 #26
