+++ BE/airdnb/src/main/java/com/team01/airdnb/accommadation/AccommodationController.java
@@ -5,21 +5,18 @@ import com.team01.airdnb.accommadation.dto.AccommodationDetailResponse;
 import com.team01.airdnb.accommadation.dto.AccommodationFilterRequest;
 import com.team01.airdnb.accommadation.dto.AccommodationRegisterRequest;
 import com.team01.airdnb.accommadation.dto.AccommodationSearchResponse;
-import com.team01.airdnb.accommadation.dto.AccommodationUpdateRequest;
 import java.util.List;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/accommodations")
 public class AccommodationController {
 
   AccommodationService accommodationService;
@@ -30,39 +27,21 @@ public class AccommodationController {
   }
 
   //숙소 생성
-  @PostMapping("/accommodations")
+  @PostMapping
   public void registerAccommodation(
       @RequestBody AccommodationRegisterRequest accommodationRegisterRequest) {
     accommodationService.register(accommodationRegisterRequest);
   }
 
-  //숙소 검색
-  @PostMapping("/accommodations/search")
-  public void searchAccommodations() {
-
-  }
-
   //숙소 상세보기
-  @GetMapping("/accommodations/{id}")
+  @GetMapping("/{id}")
   public AccommodationDetailResponse showAccommodation(@PathVariable Long id) {
     return accommodationService.show(id);
   }
 
-  //숙소 수정
-  @PutMapping("/accommodations/{id}")
-  public void updateAccommodation(@PathVariable Long id,
-      @RequestBody AccommodationUpdateRequest accommodationUpdateRequest) {
-    accommodationService.update(id, accommodationUpdateRequest);
-  }
-
-  //숙소 삭제
-  @DeleteMapping("accommodations/{id}")
-  public void deleteAccommodation(@PathVariable Long id) {
-    accommodationService.delete(id);
-  }
-
-  @GetMapping("accommodations/filter")
-  public List<AccommodationSearchResponse> getAccommodationFilter(
+  //숙소 필터
+  @GetMapping("/filter")
+  public List<AccommodationSearchResponse> getFilteredAccommodations(
       @ModelAttribute AccommodationFilterRequest accommodationFilterRequest) {
     return accommodationService.searchFilteredAccommodations(accommodationFilterRequest);
   }
